feat: adicionando service, mudando nome de arquivo, removendo metodos do plural e outras pequenas modificacoes para testes
